requirements.txt
rapidfuzz>=3.0
python-Levenshtein>=0.27
sentence-transformers>=4.0
qdrant-client>=1.10
gradio>=4.0
openai>=1.40.0
python-dotenv>=1.0.0
pymupdf>=1.24
requests>=2.28
